+++ utils/values.ts
@@ -1,4 +1,4 @@
-import { SpotlightLocation, formatNumber } from '.';
+import { SpotlightLocation } from '.';
 
 export interface ValueOptions {
   useLocalValue?: boolean;
@@ -14,16 +14,4 @@ export const DEFAULT_VALUE = 'No Data';
 
 export const addPrefixAndSuffix = (value: string | number, options: ValueOptions): string => {
   return `${options.prefix || ''} ${value}${options.suffix || ''}`;
-};
-export const formatSeries = (
-  index: number,
-  name: string | undefined,
-  seriesName: string | undefined,
-  value: number,
-  valueOptions: ValueOptions[]
-): string => {
-  return `<div>${name}<ul><li>${seriesName}: ${addPrefixAndSuffix(
-    formatNumber(value, 1),
-    valueOptions[index]
-  )}</li></ul></div>`;
 };
